fix: remove timeLeft and add hover detection to whole crop
