+++ src/features/crops/components/Soil.tsx
@@ -1,9 +1,9 @@
-import React, { useState } from "react";
+import React from "react";
 
 import classNames from "classnames";
 import soil from "assets/land/soil2.png";
 
-import { getTimeLeft, secondsToLongString } from "lib/utils/time";
+import { getTimeLeft } from "lib/utils/time";
 
 import { ProgressBar } from "components/ui/ProgressBar";
 import { InnerPanel, Panel } from "components/ui/Panel";
@@ -16,14 +16,14 @@ import classnames from "classnames";
 interface Props {
   field?: FieldItem;
   className?: string;
+  showCropDetails?: Boolean;
 }
 
-export const Soil: React.FC<Props> = ({ field, className }) => {
+export const Soil: React.FC<Props> = ({ field, className, showCropDetails }) => {
   const [_, setTimer] = React.useState<number>(0);
   const setHarvestTime = React.useCallback(() => {
     setTimer((count) => count + 1);
   }, []);
-  const [showCropDetails, setShowCropDetails] = useState(false);
 
   React.useEffect(() => {
     if (field) {
@@ -40,26 +40,13 @@ export const Soil: React.FC<Props> = ({ field, className }) => {
   const crop = CROPS()[field.name];
   const lifecycle = LIFECYCLE[field.name];
   const timeLeft = getTimeLeft(field.plantedAt, crop.harvestSeconds);
-  const timeLeftString = secondsToLongString(timeLeft);
-
-  const handleMouseHover = () => {
-    setShowCropDetails(true);
-  };
-
-  const handleMouseLeave = () => {
-    setShowCropDetails(false);
-  };
 
   // Seedling
   if (timeLeft > 0) {
     const percentage = 100 - (timeLeft / crop.harvestSeconds) * 100;
 
     return (
-      <div
-        className="relative w-full h-full"
-        onMouseEnter={handleMouseHover}
-        onMouseLeave={handleMouseLeave}
-      >
+      <div className="relative w-full h-full">
         <img
           src={lifecycle.seedling}
           className={classnames("w-full", className)}
@@ -69,15 +56,17 @@ export const Soil: React.FC<Props> = ({ field, className }) => {
         </div>
         <InnerPanel
           className={classNames(
-            "text-xxs text-white text-shadow ml-10 transition-opacity absolute whitespace-nowrap sm:opacity-0 bottom-1 w-fit left-1 z-20 pointer-events-none",
+            "ml-10 transition-opacity absolute whitespace-nowrap sm:opacity-0 bottom-5 w-fit left-1 z-20 pointer-events-none",
             {
               "opacity-100": showCropDetails,
               "opacity-0": !showCropDetails,
             }
           )}
         >
-          {field.name} <br />
-          {timeLeftString}
+          <div className="flex items-center justify-center text-xxs text-white text-shadow ml-2 mr-2">
+            <img src={lifecycle.ready} className="w-4 mr-1" />
+            <span className="flex-1">{field.name}</span>
+          </div>
         </InnerPanel>
       </div>
     );
